refactor: moved error types and python bindings into their own modules
